+++ quickbooks/objects/creditcardpayment_entity.py
@@ -1,5 +1,5 @@
 from six import python_2_unicode_compatible
-from .base import Ref, QuickbooksManagedObject, QuickbooksTransactionEntity, LinkedTxnMixin
+from .base import Ref, QuickbooksManagedObject, QuickbooksTransactionEntity, LinkedTxnMixin, MetaData
 from ..mixins import DeleteMixin
 
 
@@ -17,6 +17,8 @@ class CreditCardPayment(DeleteMixin, QuickbooksManagedObject, QuickbooksTransact
     class_dict = {
         "BankAccountRef": Ref,
         "CreditCardAccountRef": Ref,
+        "VendorRef": Ref,
+        "MetaData": MetaData,
     }
 
     qbo_object_name = "CreditCardPayment"
@@ -27,9 +29,14 @@ class CreditCardPayment(DeleteMixin, QuickbooksManagedObject, QuickbooksTransact
         self.TxnDate = None
         self.Amount = 0
         self.PrivateNote = None
+        self.Memo = None
+        self.PrintStatus = None
+        self.CheckNum = None
 
         self.BankAccountRef = None
         self.CreditCardAccountRef = None
+        self.VendorRef = None
+        self.MetaData = None
 
     def __str__(self):
         return str(self.Amount)